+++ src/main/java/com/atlauncher/gui/dialogs/LoginWithMicrosoftDialog.java
@@ -62,7 +62,7 @@ public final class LoginWithMicrosoftDialog extends JDialog {
     private static final HTTPServer server = new HTTPServer(Constants.MICROSOFT_LOGIN_REDIRECT_PORT);
     private static final VirtualHost host = server.getVirtualHost(null);
 
-    private MicrosoftAccount account = null;
+    public MicrosoftAccount account = null;
 
     public LoginWithMicrosoftDialog() {
         this(null);
@@ -192,6 +192,7 @@ public final class LoginWithMicrosoftDialog extends JDialog {
                     profile);
 
             AccountManager.addAccount(account);
+            this.account = account;
         }
     }